+++ src/SFA.DAS.FAT.Web.UnitTests/Models/WhenCreatingCoursesViewModel.cs
@@ -24,50 +24,6 @@ namespace SFA.DAS.FAT.Web.UnitTests.Models
         }
         
         [Test, AutoData]
-        public void Then_The_Clear_Filter_Items_Are_Built_From_The_Selected_Sectors(List<Guid> selectedRoutes, string keyword)
-        {
-            //Arrange Act
-            var model = CoursesViewModelTests.CoursesViewModelFactory.BuildModel(selectedRoutes, keyword, new List<int>());
-
-            //Assert
-            var clearLinkCount = selectedRoutes.Count;
-            Assert.AreEqual(clearLinkCount,model.ClearSectorLinks.Count);
-            
-            foreach (var selectedRoute in selectedRoutes)
-            {
-                var sector = model.Sectors.SingleOrDefault(c => c.Id.Equals(selectedRoute));
-                
-                Assert.IsNotNull(sector);
-                Assert.IsTrue(model.ClearSectorLinks.ContainsKey(sector.Route));
-                Assert.AreEqual(clearLinkCount-1, model.ClearSectorLinks.Count(c=>c.Value.Contains($"sectors={selectedRoute}")));    
-                Assert.AreEqual(clearLinkCount, model.ClearSectorLinks.Count(c=>c.Value.Contains($"keyword={keyword}")));    
-            }
-
-            
-        }
-
-        [Test, AutoData]
-        public void Then_The_Clear_Filter_Items_Are_Built_From_The_Selected_Levels(List<int> selectedLevels, string keyword)
-        {
-            //Arrange Act
-            var model = CoursesViewModelTests.CoursesViewModelFactory.BuildModel(new List<Guid>(), keyword, selectedLevels);
-
-            //Assert
-            var clearLinkCount = selectedLevels.Count;
-            Assert.AreEqual(clearLinkCount,model.ClearLevelLinks.Count);
-            
-            foreach (var selectedLevel in selectedLevels)
-            {
-                var levelViewModel = model.Levels.SingleOrDefault(c => c.Code.Equals(selectedLevel));
-                
-                Assert.IsNotNull(levelViewModel);
-                Assert.IsTrue(model.ClearLevelLinks.ContainsKey(levelViewModel.Title));
-                Assert.AreEqual(clearLinkCount-1, model.ClearLevelLinks.Count(c=>c.Value.Contains($"levels={selectedLevel}")));    
-                Assert.AreEqual(clearLinkCount, model.ClearLevelLinks.Count(c=>c.Value.Contains($"keyword={keyword}")));    
-            }
-        }
-        
-        [Test, AutoData]
         public void Then_The_Clear_Filter_Items_Are_Built_From_The_Selected_Levels_And_Sectors(List<int> selectedLevels, List<Guid> selectedRoutes, string keyword)
         {
             //Arrange Act
